requirements.txt
django
django-debug-toolbar
django-rest-framework
djangorestframework-simplejwt